bpf: compress countmin state to 64 bit
Use 24 instead of 32 bits to store the last observed rate. This means we
cannot measure rates above 16m pps, but in practice that is unlikely to be
a problem.

Use 40 instead of 64bits to store the timestamp of the last observed rate.
This will increase the rate at which a timestamp wrap from once every
584 years (aka never) to once every 18 minutes. This means we now have to
consider what happens when such an overflow happens.